requirements.txt
asgiref==3.2.7
Django==3.0.7
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
